+++ CHANGELOG.md
@@ -21,6 +21,8 @@
  - Added identify device extension to metadata page.
  - Added standby (motors on, laser off) state extension to metadata page.
  - Added a setting to use standby instead of sleep by default.
+ - Added sleep state extension to metadata page.
+ - Added on state extension to metadata page.
 
 # Version 1.0.1+3 2020-08-31